Attempt to detect files with the wrong extension
It is fairly common to encounter JPEG images that have a HEIC file
extension.
The plugin will now check for the file signatures of a few common
formats when the HEIC 'ftyp' box is not present.

If a supported image format is detected, the image will be loaded
using GDI+.

Closes #4