Refactor the code that uses SafeHeifImageHandle
The SafeHeifImageHandle and its associated image information are now
encapsulated in a HeifImageHandle class.
This simplifies the code in HeicLoad.cs.

PrimaryImageInfo has been renamed to ImageHandleInfo, and now includes
the CICP color data.

The GetCICPColorData function has been replaced by a GetMetadataId
function, which eliminates the double id lookup when reading EXIF and
XMP metadata.