Improve loading of Rec. 2020 PQ HDR images
These images are now loaded as Float32 and converted to Bgra32 using
Direct2D, similar to the AVIF FileType.
The conversion of SDR images to Bgra32 should also be improved by using
WIC instead of libheif.

Added code to check if the loaded ICC profile is RGB, and map some CICP
values to the matching Paint.NET color profile.